feat: open white dwarf modal from subtype; move dialog to shell
Remove the cyan spectral-code badge and make the body subtype itself the
click target that opens the White Dwarf Spectral Types modal.

Move the dialog into src/app/dialogs/ and reuse DialogShellComponent for
its title/content/close chrome, matching the H-R diagram dialog pattern.

Update the data-driven e2e spec to click the subtype (the badge is gone)
and add a Phoi Scraa HH-V f2-67 fixture/case (a DA white dwarf).
