+++ e2e/fixtures/phoi-scraa-hh-v-f2-67.json
@@ -1,0 +1,1 @@
+{"system":{"allegiance":null,"bodies":[{"absoluteMagnitude":20.0,"age":8420,"argOfPeriapsis":304.810541,"axialTilt":0.0,"bodyId":1,"distanceToArrival":0.0,"id64":36028833125135549,"luminosity":"VII","mainStar":true,"name":"Phoi Scraa HH-V f2-67 A","orbitalEccentricity":0.260771,"orbitalInclination":91.15927,"orbitalPeriod":2016762.75443148,"parents":[{"Null":0}],"rotationalPeriod":1.56081018518519e-06,"semiMajorAxis":46.6069822065426,"solarMasses":3.90625,"solarRadius":1.65684337929547e-05,"spectralClass":null,"stations":[],"subType":"Black Hole","surfaceTemperature":0.0,"type":"Star","updateTime":"2021-03-02 12:46:58+00"},{"absoluteMagnitude":9.76532,"age":8420,"argOfPeriapsis":124.810547,"axialTilt":0.0,"bodyId":2,"distanceToArrival":309072.0,"id64":72057630144099517,"luminosity":"VII","mainStar":false,"name":"Phoi Scraa HH-V f2-67 B","orbitalEccentricity":0.260771,"orbitalInclination":91.15927,"orbitalPeriod":2016762.75443148,"parents":[{"Null":0}],"rotationalPeriod":0.946201657662037,"rotationalPeriodTidallyLocked":false,"semiMajorAxis":461.455297064733,"solarMasses":0.394531,"solarRadius":0.0184631329978433,"stations":[],"subType":"White Dwarf (DA) Star","surfaceTemperature":13653.0,"type":"Star","updateTime":"2021-03-02 12:47:00+00"}],"bodyCount":2,"coords":{"x":-8461.65625,"y":2248.625,"z":24517.65625},"date":"2021-03-02 12:46:51+00","government":"None","id64":36106171581,"name":"Phoi Scraa HH-V f2-67","population":0,"primaryEconomy":"None","region":{"name":"Odin's Hold","region":4},"secondaryEconomy":"None","security":"Anarchy","stations":[]}}
